Centraliser les gestions de flashmessages
