+++ inertia/app/app.tsx
@@ -6,11 +6,12 @@ import '@codegouvfr/react-dsfr/dsfr/dsfr.css'
 import '@codegouvfr/react-dsfr/dsfr/utility/icons/icons.css'
 import '../css/app.css'
 import { hydrateRoot } from 'react-dom/client'
-import { createInertiaApp, Link } from '@inertiajs/react'
+import { createInertiaApp, Link, router } from '@inertiajs/react'
 import { resolvePageComponent } from '@adonisjs/inertia/helpers'
 import { startReactDsfr } from '@codegouvfr/react-dsfr/spa'
 import { fr } from '@codegouvfr/react-dsfr'
-import { Toaster } from '../ui/Toaster'
+import { toast } from 'react-toastify'
+import Toast, { Toaster } from '../ui/Toaster'
 
 startReactDsfr({ defaultColorScheme: 'system', Link })
 declare module '@codegouvfr/react-dsfr/spa' {
@@ -22,6 +23,39 @@ declare module '@codegouvfr/react-dsfr/spa' {
 const appName = import.meta.env.VITE_APP_NAME || 'Viz’Eau'
 
 const blueFrance = fr.colors.decisions.background.actionHigh.blueFrance.active
+const TOAST_SEVERITIES = ['success', 'error', 'warning', 'info'] as const
+
+type ToastSeverity = (typeof TOAST_SEVERITIES)[number]
+
+type FlashMessages = Partial<
+  Record<
+    ToastSeverity,
+    {
+      message?: string
+      context?: string
+    } | null
+  >
+>
+
+function showFlashToasts(flashMessages?: FlashMessages) {
+  const alerts = TOAST_SEVERITIES.filter(
+    (severity) => flashMessages?.[severity]?.message && !flashMessages?.[severity]?.context
+  ).map((severity) => ({
+    severity,
+    message: flashMessages?.[severity]?.message as string,
+  }))
+
+  if (alerts.length === 0) return
+
+  const toastId = alerts.map((a) => `${a.severity}:${a.message}`).join('|')
+  if (toast.isActive(toastId)) return
+
+  toast(<Toast alerts={alerts} />, {
+    toastId,
+    closeButton: false,
+    style: { padding: 0, background: 'transparent', boxShadow: 'none' },
+  })
+}
 
 createInertiaApp({
   progress: { color: blueFrance },
@@ -33,6 +67,12 @@ createInertiaApp({
   },
 
   setup({ el, App, props }) {
+    showFlashToasts(props.initialPage.props.flashMessages as FlashMessages | undefined)
+
+    router.on('success', (event) => {
+      showFlashToasts(event.detail.page.props.flashMessages as FlashMessages | undefined)
+    })
+
     hydrateRoot(
       el,
       <>
