+++ inertia/components/parcelles-manager.tsx
@@ -1,9 +1,7 @@
 import Button from '@codegouvfr/react-dsfr/Button'
 import { createModal } from '@codegouvfr/react-dsfr/Modal'
 import { fr } from '@codegouvfr/react-dsfr'
-import { toast } from 'react-toastify'
 import Loader from '~/ui/Loader'
-import Toast from '~/ui/Toaster'
 import { ExploitationJson } from '../../types/models'
 
 export type ParcellesManagerProps = {
@@ -37,19 +35,6 @@ export default function ParcellesManager({
     id: 'cancel-edit-mode-modal',
     isOpenedByDefault: false,
   })
-  const showSuccessToast = () => {
-    toast(
-      <Toast
-        alerts={[
-          { severity: 'success', message: 'Les parcelles ont été mises à jour avec succès.' },
-        ]}
-      />,
-      {
-        closeButton: false,
-        style: { padding: 0, background: 'transparent', boxShadow: 'none' },
-      }
-    )
-  }
 
   return (
     <div
@@ -88,7 +73,7 @@ export default function ParcellesManager({
               size="small"
               disabled={processing}
               onClick={() => {
-                sendFormAndResetState(showSuccessToast)
+                sendFormAndResetState()
               }}
             >
               Appliquer
@@ -134,7 +119,7 @@ export default function ParcellesManager({
             children: 'Appliquer les modifications',
             disabled: processing,
             onClick: () => {
-              sendFormAndResetState(showSuccessToast)
+              sendFormAndResetState()
             },
           },
         ]}
